moved dietary restrictions implementation from app.jsx to pages/dietaryrestrictions.js
